MAINT: Move form file to subdirectory
